+++ src/main/java/com/ats/tril/model/StockDetail.java
@@ -23,28 +23,28 @@ public class StockDetail {
 	private int itemId;
 
 	@Column(name = "op_stock_qty")
-	private int opStockQty;
+	private float opStockQty;
 
 	@Column(name = "approved_qty")
-	private int approvedQty;
+	private float approvedQty;
 	
 	@Column(name = "issue_qty")
-	private int issueQty;
+	private float issueQty;
 	
 	@Column(name = "return_issue_qty")
-	private int returnIssueQty;
+	private float returnIssueQty;
 	
 	@Column(name = "damage_qty")
-	private int damageQty;
+	private float damageQty;
 	
 	@Column(name = "gatepass_qty")
-	private int gatepassQty;
+	private float gatepassQty;
 	
 	@Column(name = "gatepass_return_qty")
-	private int gatepassReturnQty;
+	private float gatepassReturnQty;
 	
 	@Column(name = "closing_qty")
-	private int closingQty;
+	private float closingQty;
 	
 	@Column(name = "del_status")
 	private int delStatus;
@@ -73,67 +73,67 @@ public class StockDetail {
 		this.itemId = itemId;
 	}
 
-	public int getOpStockQty() {
+	public float getOpStockQty() {
 		return opStockQty;
 	}
 
-	public void setOpStockQty(int opStockQty) {
+	public void setOpStockQty(float opStockQty) {
 		this.opStockQty = opStockQty;
 	}
 
-	public int getApprovedQty() {
+	public float getApprovedQty() {
 		return approvedQty;
 	}
 
-	public void setApprovedQty(int approvedQty) {
+	public void setApprovedQty(float approvedQty) {
 		this.approvedQty = approvedQty;
 	}
 
-	public int getIssueQty() {
+	public float getIssueQty() {
 		return issueQty;
 	}
 
-	public void setIssueQty(int issueQty) {
+	public void setIssueQty(float issueQty) {
 		this.issueQty = issueQty;
 	}
 
-	public int getReturnIssueQty() {
+	public float getReturnIssueQty() {
 		return returnIssueQty;
 	}
 
-	public void setReturnIssueQty(int returnIssueQty) {
+	public void setReturnIssueQty(float returnIssueQty) {
 		this.returnIssueQty = returnIssueQty;
 	}
 
-	public int getDamageQty() {
+	public float getDamageQty() {
 		return damageQty;
 	}
 
-	public void setDamageQty(int damageQty) {
+	public void setDamageQty(float damageQty) {
 		this.damageQty = damageQty;
 	}
 
-	public int getGatepassQty() {
+	public float getGatepassQty() {
 		return gatepassQty;
 	}
 
-	public void setGatepassQty(int gatepassQty) {
+	public void setGatepassQty(float gatepassQty) {
 		this.gatepassQty = gatepassQty;
 	}
 
-	public int getGatepassReturnQty() {
+	public float getGatepassReturnQty() {
 		return gatepassReturnQty;
 	}
 
-	public void setGatepassReturnQty(int gatepassReturnQty) {
+	public void setGatepassReturnQty(float gatepassReturnQty) {
 		this.gatepassReturnQty = gatepassReturnQty;
 	}
 
-	public int getClosingQty() {
+	public float getClosingQty() {
 		return closingQty;
 	}
 
-	public void setClosingQty(int closingQty) {
+	public void setClosingQty(float closingQty) {
 		this.closingQty = closingQty;
 	}
